aethercore-backend/requirements.txt
fastapi
uvicorn
auth0-python
python-jose
httpx
pydantic
pydantic-settings
sqlalchemy>=2.0,<3.0
alembic>=1.13,<2.0
psycopg[binary]>=3.1,<4.0
celery>=5.4,<6.0
redis>=5.0,<6.0
prometheus-client>=0.21,<1.0
opentelemetry-api>=1.28,<2.0
opentelemetry-sdk>=1.28,<2.0
opentelemetry-exporter-otlp>=1.28,<2.0
opentelemetry-instrumentation-fastapi>=0.49b0
opentelemetry-instrumentation-httpx>=0.49b0
pytest>=7.4,<8.0
pytest-asyncio>=0.23,<1.0
pytest-cov>=4.1,<5.0
httpx>=0.25,<1.0
mypy>=1.8,<2.0
ruff>=0.3,<1.0